Converted dailyCharge from float to BigDecimal for precise computation.

+++ src/main/java/com/nm1022/tool/rental/persistence/model/RentalCharge.java
@@ -2,15 +2,17 @@ package com.nm1022.tool.rental.persistence.model;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class RentalCharge {
 
-    private float dailyCharge;
+    private BigDecimal dailyCharge;
     private ChargeIndicators chargeIndicators;
 
     private RentalCharge() { };
 
-    public RentalCharge(float dailyCharge, ChargeIndicators chargeIndicators) {
+    public RentalCharge(BigDecimal dailyCharge, ChargeIndicators chargeIndicators) {
         this.dailyCharge = dailyCharge;
         this.chargeIndicators = chargeIndicators;
     }
